requirements.txt
git+https://github.com/biden2020prez/groupme-push
aiopyston==1.2.1
alt_profanity_check==1.3.2
colorama==0.4.6
dataset==1.6.2
Flask==1.1.1
groupme_push==0.0.3
profanity_check==1.0.3
Pygments==2.17.2
python-dotenv==1.0.1
regex_spm==1.0.0
Requests==2.31.0
websockets==12.0
psycopg2==2.9.9